modules/CustomVisionContainer/requirements.txt
requests
Pillow
Flask
numpy